# api-gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GIGA API Gateway",
    "version": "2.0.0",
...
    "contact": {
      "name": "GIGA Platform Team",
      "email": "api@giga.com"
    }
  },
  "servers": [
    {
      "url": "https://giga-giga-production.up.railway.app",
      "description": "Production API Gateway"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Supabase Auth"
      }
    },
    "parameters": {
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "description": "Page number"
      },
      "Limit": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        },
        "description": "Items per page"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/social/posts": {
      "post": {
        "summary": "Create a post",
        "description": "Create a new social media post",
        "tags": ["Social Media"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "media_urls": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["public", "friends", "private"],
                    "default": "public"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/social/posts/{id}": {
      "get": {
        "summary": "Get post details",
        "description": "Get detailed information about a specific post",
        "tags": ["Social Media"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post details"
          }
        }
      },
      "put": {
        "summary": "Update post",
        "description": "Update an existing post",
        "tags": ["Social Media"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete post",
        "description": "Delete a post",
        "tags": ["Social Media"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/social/users/{id}/posts": {
      "get": {
        "summary": "Get user posts",
        "description": "Get posts by a specific user",
        "tags": ["Social Media"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "User's posts"
          }
        }
      }
    },
...
  }
}